+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2021, Wang Lab at Tongji University'
 author = 'Xin Dong'
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.20211219'
+release = '0.1.240527'
 
 
 # -- General configuration ---------------------------------------------------