+++ docs/release_notes.rst
@@ -1,7 +1,15 @@
 Release notes
 ===================================================
 
-Version 0.1.220630  *2021-06-30*
+Version 0.1.240527 *2024-05-27*
+---------------------------------
+
+- Remove GIGGLE built-in and add the function and documentation for installing GIGGLE .
+- Support MTX input for impute function.
+- Documents improvements.
+
+
+Version 0.1.220630  *2022-06-30*
 ---------------------------------
 
 - Add index function to SCRIP. Users can create a self-defined index.